Move IR test files into folder as well
