+++ quint/test/internal_representation/IRFinder.test.ts
@@ -1,7 +1,7 @@
 import { assert } from 'chai'
 import { describe, it } from 'mocha'
-import { findDefinitionWithId, findExpressionWithId, findTypeWithId } from '../src/internal_representation/IRFinder'
-import { buildModuleWithDefs, buildModuleWithExpressions } from './builders/ir'
+import { findDefinitionWithId, findExpressionWithId, findTypeWithId } from '../../src/internal_representation/IRFinder'
+import { buildModuleWithDefs, buildModuleWithExpressions } from '../builders/ir'
 
 describe('findExpressionWithId', () => {
   const modules = [buildModuleWithExpressions(['Nat'])]
