+++ quint/test/internal_representation/quintTypes.test.ts
@@ -1,7 +1,7 @@
 import { describe, it } from 'mocha'
 import { assert } from 'chai'
-import { typeNames } from '../src/internal_representation/quintTypes'
-import { parseType, parseTypeOrThrow } from '../src/types/parser'
+import { typeNames } from '../../src/internal_representation/quintTypes'
+import { parseType, parseTypeOrThrow } from '../../src/types/parser'
 
 describe('typeNames', () => {
   it('find names in types', () => {
